simple size dependencies added
It is actually not quite trivial to allow for linked regions and other
fun if you want to let the compiler figure out what to do.

The solution, for now, is to have constraint kinds, and allow for
arbitrary constraints based on type classes.

At least for the smallish example I am using now, this leads to
efficient code (funnily enough). Now blowing up to something real ...