this version descends the index stack
again, we use constraint kinds to allow dynamic class constraints.
dynamic as in the user can write constraints and apply them to the index
stack. the compiler compiles into efficient code

note that this is the ``tech demo'', not working code as in working on
the level of CFGs or anything ...